Print only one line at the end of each test
We used to print the start of the test message, run the test, then finish it.
The parallel test suite tended to mess this up and generate a bunch of weird
output.

We give up knowing what test broke if a test has hung, but the quality of that
information was already not great. Again this is thanks to the parallel test
framework: the test that hung is not necessarily last to speak.

Signed-off-by: Casey Dahlin <cdahlin@redhat.com>