Introduce ason_reduce and make it core
ason_reduce reduces the form of an ASON value in place. It's goal is to
determine whether a value is the empty value, and it returns a flag as to
whether it is.

All of the comparison operations can be implemented in terms of reducing some
expression and checking for whether the result is empty, so we now do this.

This operation replaces ason_simplify_transform, ason_flatten, etc. Its purpose
is more concrete as well.

Signed-off-by: Casey Dahlin <cdahlin@redhat.com>