Change blockfile API
We now use a custom bfsize_t instead of size_t, which is guaranteed to be 64
bits on all platforms. We also have a block_t type which is used when we're
passing block numbers, though it is the same dimensions as bfsize_t.

We then introduce several error constant block numbers at the end of the value
range, so we never need to have a signed bfsize_t or block_t.

Signed-off-by: Casey Dahlin <cdahlin@redhat.com>