Add B-Tree implementation
This code implements a B-tree on top of block file blocks. It's compile-tested
only right now, but it needs to be in source control before it gets any
hairier.

Nodes are one block long, and contain key-value pairs. The first key is
ignored since B-trees usually have one fewer key than children, but the first
key still must be the least key, because we store the keys unsorted, which
saves disk syncs and helps us stay atomic, but shifts more load onto the CPU.
Time will tell whether this is the right decision.

Each B-tree has a metadata "head block" which contains, among other things, a
journal. The journal gets slightly less than a block's worth of space
(whatever's left after the B-tree meta data and root pointer). The journal size
bounds how complex a restructuring operation we can perform on the tree, and
thus ultimately the height of the tree. I haven't computed the bound but it
should be large enough. It's format is simple enough: it records changes to the
tree such as adding keys, removing them, merging blocks, etc. When it's time to
commit, we add a checksum trailer and begin executing the changes.

Signed-off-by: Casey Dahlin <cdahlin@redhat.com>