Rewrite ason_reduce_intersect to use orders
We can reduce intersection much more cleanly now. And, indeed, the
recently-failing test cases now pass.

We've uncovered a category of defects in intersection that still remain (we
largely don't handle order 3 values correctly) but it's a more defined problem
now, and the cases it occurs in are more likely to be rare.

Signed-off-by: Casey Dahlin <casey.dahlin@gmail.com>