Final representation check based on complement of right operand
We used to base our last-ditch representation check on the intersection of the
two operands and whether it was equal to the left operand. Now we base it on
the intersection of the left operand and the complement of the right operand,
which should be empty if the test is true. Since in this last ditch case the
right operand is always a complement, we are removing complexity and thus are
less likely to be inducing a loop.

Signed-off-by: Casey Dahlin <casey.dahlin@gmail.com>