Initial code for number domains
I've convinced myself that it's safe to treat an ASON value as a union of ASON
values which represent values of only one primitive type, and to perform all
operations by operating on these "domain" values as individual pairs and
combining the results. This will lead to much simpler (and faster) logic in the
long run, but it's pretty incompatible with our current code.

This commit introduces a structure for storing just integers. All operations on
it should be O(n). It's not yet hooked up to anything (indeed it even lacks
tests), but it's a start.

Signed-off-by: Casey Dahlin <casey.dahlin@gmail.com>