Begin transition to domain-based ASON computation
There's sort of a big mess of changes here, but they're hard to extract. Also
this totally breaks 2/3 of the test suite, but for good reason.

So, previously ASON stored values internally as what were essentially ASTs,
which it would transform into a reduced normal form to perform computation etc.

I several times had the idea of making domained unions the primitive. This
would mean each internally stored node would be a union of values (a "union" of
just one value if necessary), and the union would be separated into type
domains, so a collection of numbers, a collection of strings, a collection of
objects, etc. Operating on two of these unioned sets would usually mean
operating on the individual collections in turn. I.e. to intersect two domained
unions, you intersect the integer portion, then the string portion, etc. The
intersections on each of those domains is much easier to define then
intersection on a general ASON value, so this would greatly simplify the
implementation and improve correctness. It's also likely to enable quicker
wins for performance.

I've been back and forth on this idea for a long time, mostly because I
couldn't reliably convince myself it didn't break correctness to decompose the
operations that way, particularly for the join operator. However, after further
deliberation, I'm feeling confident in it.

This commit removes all of the old reduction logic, and stumpifies many of the
functions that can't be quite implemented without it. It also implements a
domain for numbers only (more to come of course).

Obviously this means the code doesn't work right anymore, but it's a good start
to getting a livable permanent implementation that isn't broken anywhere
(there's still a corner case of join that the previous code never handled).
A lot of the planned features are also now trivial (in the number
domain implementation, we basically grew support for continuous ranges like
">=3" for free).

Signed-off-by: Casey Dahlin <casey.dahlin@gmail.com>